+++ app/src/main/java/fr/jerome/climbinggymlog/models/Personne.java
@@ -6,7 +6,7 @@ package fr.jerome.climbinggymlog.models;
  */
 public abstract class Personne {
 
-    protected long id;
+    protected int id;
     protected String nom;
     protected String prenom;
     protected int age;
@@ -20,7 +20,7 @@ public abstract class Personne {
      * @param prenom de la personne
      * @param age de la personne
      */
-    public Personne(long id, String nom, String prenom, int age, String email) {
+    public Personne(int id, String nom, String prenom, int age, String email) {
 
         this.id = id;
         this.nom = nom;
@@ -51,43 +51,35 @@ public abstract class Personne {
 
     }
 
-    public long getId() {
-
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
-
+    public void setId(int id) {
         this.id = id;
     }
 
     public String getNom() {
-
         return nom;
     }
 
     public void setNom(String nom) {
-
         this.nom = nom;
     }
 
     public String getPrenom() {
-
         return prenom;
     }
 
     public void setPrenom(String prenom) {
-
         this.prenom = prenom;
     }
 
     public int getAge() {
-
         return age;
     }
 
     public void setAge(int age) {
-
         this.age = age;
     }
